@@ -11,6 +11,9 @@ import csv
 import pytest
 import zipfile
 import random
+import warnings
+from pathlib import Path
+
 import numpy
 
 import sourmash_tst_utils as utils
@@ -330,6 +333,15 @@ def test_compare_quiet(runtmp):
                    'short2.fa.sig', '--csv', 'xxx', '-q')
     assert not c.last_result.out
     assert not c.last_result.err
+
+
+def test_compare_do_traverse_directory_parse_args(runtmp):
+    # test 'compare' on a directory, using sourmash.cli.parse_args.
+    import sourmash.commands, sourmash.cli
+    args = sourmash.cli.parse_args(['compare', '-k', '21', '--dna',
+                                    utils.get_test_data('compare')])
+
+    sourmash.commands.compare(args)
 
 
 def test_compare_do_traverse_directory(runtmp):
@@ -829,8 +841,10 @@ def test_do_plot_comparison(runtmp):
     assert os.path.exists(c.output("cmp.matrix.png"))
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_2(c):
+def test_do_plot_comparison_2_pdf(runtmp):
+    # test plot --pdf
+    c = runtmp
+
     testdata1 = utils.get_test_data('short.fa')
     testdata2 = utils.get_test_data('short2.fa')
     c.run_sourmash('sketch', 'translate', '-p', 'k=31,num=500', testdata1, testdata2)
@@ -842,8 +856,10 @@ def test_do_plot_comparison_2(c):
     assert os.path.exists(c.output("cmp.matrix.pdf"))
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_3(c):
+def test_do_plot_comparison_3(runtmp):
+    # test plot --labels
+    c = runtmp
+
     testdata1 = utils.get_test_data('short.fa')
     testdata2 = utils.get_test_data('short2.fa')
     c.run_sourmash('sketch', 'translate', '-p', 'k=31,num=500', testdata1, testdata2)
@@ -856,8 +872,10 @@ def test_do_plot_comparison_3(c):
     assert os.path.exists(c.output("cmp.matrix.png"))
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_4_output_dir(c):
+def test_do_plot_comparison_4_output_dir(runtmp):
+    # test plot --output-dir
+    c = runtmp
+
     output_dir = c.output('xyz_test')
 
     testdata1 = utils.get_test_data('short.fa')
@@ -872,8 +890,10 @@ def test_do_plot_comparison_4_output_dir(c):
     assert os.path.exists(os.path.join(output_dir, "cmp.matrix.png"))
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_5_force(c):
+def test_do_plot_comparison_5_force(runtmp):
+    # test -f to force display of something that's not a distance matrix
+    c = runtmp
+
     D = numpy.zeros([2, 2])
     D[0, 0] = 5
     with open(c.output('cmp'), 'wb') as fp:
@@ -887,8 +907,10 @@ def test_do_plot_comparison_5_force(c):
     assert c.last_result.status == 0
 
 
-@utils.in_tempdir
-def test_do_plot_comparison_4_fail_not_distance(c):
+def test_do_plot_comparison_4_fail_not_distance(runtmp):
+    # plot should fail when not a distance matrix
+    c = runtmp
+
     D = numpy.zeros([2, 2])
     D[0, 0] = 5
     with open(c.output('cmp'), 'wb') as fp:
@@ -904,32 +926,181 @@ def test_do_plot_comparison_4_fail_not_distance(c):
     assert c.last_result.status != 0
 
 
+def test_plot_6_labels_default(runtmp):
+    # plot --labels is default
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
+
+    runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
+
+    runtmp.sourmash('plot', 'cmp', '--labels')
+
+    print(runtmp.last_result.out)
+
+    expected = """\
+0\tgenome-s10
+1\tgenome-s11
+2\tgenome-s12
+3\tgenome-s10+s11"""
+    assert expected in runtmp.last_result.out
+
+
+def test_plot_6_labels(runtmp):
+    # specifing --labels gives the right result
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
+
+    runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
+
+    runtmp.sourmash('plot', 'cmp', '--labels')
+
+    print(runtmp.last_result.out)
+
+    expected = """\
+0\tgenome-s10
+1\tgenome-s11
+2\tgenome-s12
+3\tgenome-s10+s11"""
+    assert expected in runtmp.last_result.out
+
+
+def test_plot_6_indices(runtmp):
+    # test plot --indices
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
+
+    runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
+
+    runtmp.sourmash('plot', 'cmp', '--indices')
+
+    print(runtmp.last_result.out)
+
+    expected = """\
+0\t1
+1\t2
+2\t3
+3\t4"""
+    assert expected in runtmp.last_result.out
+
+
+def test_plot_6_no_labels(runtmp):
+    # test plot --no-labels
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
+
+    runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
+
+    runtmp.sourmash('plot', 'cmp', '--no-labels')
+
+    print(runtmp.last_result.out)
+
+    expected = """\
+0\t1
+1\t2
+2\t3
+3\t4"""
+    assert expected in runtmp.last_result.out
+
+
+def test_plot_6_no_indices(runtmp):
+    # test plot --no-labels
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
+
+    runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
+
+    runtmp.sourmash('plot', 'cmp', '--no-labels')
+
+    print(runtmp.last_result.out)
+
+    expected = """\
+0\t1
+1\t2
+2\t3
+3\t4"""
+    assert expected in runtmp.last_result.out
+
+
+def test_plot_6_no_labels_no_indices(runtmp):
+    # test plot --no-labels --no-indices
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
+
+    runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
+
+    runtmp.sourmash('plot', 'cmp', '--no-labels', '--no-indices')
+
+    print((runtmp.last_result.out,))
+
+    expected = """\
+0\t
+1\t
+2\t
+3\t"""
+    assert expected in runtmp.last_result.out
+
+
+def test_plot_6_indices_labels(runtmp):
+    # check that --labels --indices => --labels
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
+
+    runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
+
+    runtmp.sourmash('plot', 'cmp', '--labels', '--indices')
+
+    print(runtmp.last_result.out)
+
+    expected = """\
+0\tgenome-s10
+1\tgenome-s11
+2\tgenome-s12
+3\tgenome-s10+s11"""
+    assert expected in runtmp.last_result.out
+
+
 def test_plot_override_labeltext(runtmp):
-        testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
-        testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
-        testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
-        testdata4 = utils.get_test_data('genome-s10+s11.sig')
+    # test overriding labeltext
+    testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
+    testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
+    testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
+    testdata4 = utils.get_test_data('genome-s10+s11.sig')
 
-        runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
+    runtmp.run_sourmash('compare', testdata1, testdata2, testdata3, testdata4, '-o', 'cmp', '-k', '21', '--dna')
 
-        with open(runtmp.output('new.labels.txt'), 'wt') as fp:
-            fp.write('a\nb\nc\nd\n')
+    with open(runtmp.output('new.labels.txt'), 'wt') as fp:
+        fp.write('a\nb\nc\nd\n')
 
-        runtmp.sourmash('plot', 'cmp', '--labeltext', 'new.labels.txt')
+    runtmp.sourmash('plot', 'cmp', '--labeltext', 'new.labels.txt')
 
-        print(runtmp.last_result.out)
+    print(runtmp.last_result.out)
 
-        assert 'loading labels from new.labels.txt' in runtmp.last_result.err
+    assert 'loading labels from new.labels.txt' in runtmp.last_result.err
 
-        expected = """\
+    expected = """\
 0\ta
 1\tb
 2\tc
 3\td"""
-        assert expected in runtmp.last_result.out
+    assert expected in runtmp.last_result.out
 
 
 def test_plot_override_labeltext_fail(runtmp):
+    # test failed override of labeltext
     testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
     testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
     testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
@@ -951,7 +1122,7 @@ def test_plot_override_labeltext_fail(runtmp):
 
 
 def test_plot_reordered_labels_csv(runtmp):
-    # test 'plot --csv'
+    # test 'plot --csv' & correct ordering of labels
     c = runtmp
 
     ss2 = utils.get_test_data('2.fa.sig')
@@ -997,6 +1168,7 @@ def test_plot_reordered_labels_csv_gz(runtmp):
 
 
 def test_plot_subsample_1(runtmp):
+    # test plotting with --subsample
     testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
     testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
     testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
@@ -1016,6 +1188,7 @@ def test_plot_subsample_1(runtmp):
 
 
 def test_plot_subsample_2(runtmp):
+    # test plotting --subsample with --subsample-seed
     testdata1 = utils.get_test_data('genome-s10.fa.gz.sig')
     testdata2 = utils.get_test_data('genome-s11.fa.gz.sig')
     testdata3 = utils.get_test_data('genome-s12.fa.gz.sig')
@@ -1488,8 +1661,7 @@ def test_do_sourmash_sbt_search_output(runtmp):
 
     runtmp.sourmash('search', 'short.fa.sig', 'zzz', '-o', 'foo')
 
-    outfile = open(runtmp.output('foo'))
-    output = outfile.read()
+    output = Path(runtmp.output('foo')).read_text()
     print(output)
     assert 'e26a306d26512' in output
     assert '914591cd1130aa915' in output
@@ -1563,8 +1735,7 @@ def test_do_sourmash_sbt_move_and_search_output(runtmp):
                                         'zzz.sbt.json', '-o', 'foo'],
                                         in_directory=newpath)
 
-    outfile = open(os.path.join(newpath, 'foo'))
-    output = outfile.read()
+    output = Path(os.path.join(newpath, 'foo')).read_text()
     print(output)
     assert '914591cd1130aa91' in output
     assert 'e26a306d2651' in output
@@ -1920,11 +2091,11 @@ def test_search_gzip(runtmp):
 
     runtmp.sourmash('sketch','dna','-p','k=31,num=500', testdata1, testdata2)
 
-    data = open(runtmp.output('short.fa.sig'), 'rb').read()
+    data = Path(runtmp.output('short.fa.sig')).read_bytes()
     with gzip.open(runtmp.output('zzz.gz'), 'wb') as fp:
         fp.write(data)
 
-    data = open(runtmp.output('short2.fa.sig'), 'rb').read()
+    data = Path(runtmp.output('short2.fa.sig')).read_bytes()
     with gzip.open(runtmp.output('yyy.gz'), 'wb') as fp:
         fp.write(data)
 
@@ -3029,10 +3200,11 @@ def test_compare_with_abundance_1(runtmp):
     s1 = signature.SourmashSignature(E1, filename='e1', name='e1')
     s2 = signature.SourmashSignature(E2, filename='e2', name='e2')
 
-    signature.save_signatures([s1],
-                                open(runtmp.output('e1.sig'), 'w'))
-    signature.save_signatures([s2],
-                                open(runtmp.output('e2.sig'), 'w'))
+    with open(runtmp.output('e1.sig'), 'w') as f:
+        signature.save_signatures([s1], f)
+
+    with open(runtmp.output('e2.sig'), 'w') as f:
+        signature.save_signatures([s2], f)
 
     runtmp.sourmash('search', 'e1.sig', 'e2.sig', '-k', '5')
 
@@ -3054,10 +3226,11 @@ def test_compare_with_abundance_2(runtmp):
     s1 = signature.SourmashSignature(E1, filename='e1', name='e1')
     s2 = signature.SourmashSignature(E2, filename='e2', name='e2')
 
-    signature.save_signatures([s1],
-                                open(runtmp.output('e1.sig'), 'w'))
-    signature.save_signatures([s2],
-                                open(runtmp.output('e2.sig'), 'w'))
+    with open(runtmp.output('e1.sig'), 'w') as f:
+        signature.save_signatures([s1], f)
+
+    with open(runtmp.output('e2.sig'), 'w') as f:
+        signature.save_signatures([s2], f)
 
     runtmp.sourmash('search', 'e1.sig', 'e2.sig', '-k', '5')
 
@@ -3080,10 +3253,11 @@ def test_compare_with_abundance_3(runtmp):
     s1 = signature.SourmashSignature(E1, filename='e1', name='e1')
     s2 = signature.SourmashSignature(E2, filename='e2', name='e2')
 
-    signature.save_signatures([s1],
-                                open(runtmp.output('e1.sig'), 'w'))
-    signature.save_signatures([s2],
-                                open(runtmp.output('e2.sig'), 'w'))
+    with open(runtmp.output('e1.sig'), 'w') as f:
+        signature.save_signatures([s1], f)
+
+    with open(runtmp.output('e2.sig'), 'w') as f:
+        signature.save_signatures([s2], f)
 
     runtmp.sourmash('search', 'e1.sig', 'e2.sig', '-k', '5')
 
@@ -5176,7 +5350,7 @@ def test_sbt_categorize(runtmp):
     # yields 521/1000 ==> ~0.5
     assert 'for genome-s10+s11, found: 0.50 genome-s10' in runtmp.last_result.err
 
-    out_csv = open(runtmp.output('out.csv')).read()
+    out_csv = Path(runtmp.output('out.csv')).read_text()
     print(out_csv)
     assert '4.sig,genome-s10+s11,genome-s10,0.504' in out_csv
 
@@ -5229,7 +5403,7 @@ def test_sbt_categorize_ignore_abundance_3(runtmp):
 
     assert 'for 1-1, found: 0.88 1-1' in runtmp.last_result.err
 
-    out_csv4 = open(runtmp.output('out4.csv')).read()
+    out_csv4 = Path(runtmp.output('out4.csv')).read_text()
     assert 'reads-s10x10-s11.sig,1-1,1-1,0.87699' in out_csv4
 
 
@@ -5617,7 +5791,7 @@ def test_do_sourmash_index_zipfile_append(c):
     # should be no overlap
     assert not set(first_half).intersection(set(second_half))
 
-    with pytest.warns(None) as record:
+    with warnings.catch_warnings(record=True) as record:
         c.run_sourmash('index', '-k', '31', 'zzz.sbt.zip',
                        *first_half)
     # UserWarning is raised when there are duplicated entries in the zipfile
@@ -5630,7 +5804,7 @@ def test_do_sourmash_index_zipfile_append(c):
     assert c.last_result.status == 0
     assert 'Finished saving SBT index, available at' in c.last_result.err
 
-    with pytest.warns(None) as record:
+    with warnings.catch_warnings(record=True) as record:
         c.run_sourmash('index', "--append", '-k', '31', 'zzz.sbt.zip',
                        *second_half)
     # UserWarning is raised when there are duplicated entries in the zipfile
