@@ -1,6 +1,6 @@
 //! # Foreign Function Interface for calling sourmash from a C API
 //!
-//! Primary client for now is the Python version, using CFFI and milksnake.
+//! Primary client for now is the Python version, using CFFI and maturin.
 #![allow(clippy::missing_safety_doc)]
 
 #[macro_use]
@@ -28,4 +28,41 @@ pub unsafe extern "C" fn hash_murmur(kmer: *const c_char, seed: u64) -> u64 {
     };
 
     _hash_murmur(c_str.to_bytes(), seed)
+}
+
+#[repr(u32)]
+pub enum HashFunctions {
+    Murmur64Dna = 1,
+    Murmur64Protein = 2,
+    Murmur64Dayhoff = 3,
+    Murmur64Hp = 4,
+}
+
+impl From<HashFunctions> for crate::encodings::HashFunctions {
+    fn from(v: HashFunctions) -> crate::encodings::HashFunctions {
+        use crate::encodings::HashFunctions::{
+            Murmur64Dayhoff, Murmur64Dna, Murmur64Hp, Murmur64Protein,
+        };
+        match v {
+            HashFunctions::Murmur64Dna => Murmur64Dna,
+            HashFunctions::Murmur64Protein => Murmur64Protein,
+            HashFunctions::Murmur64Dayhoff => Murmur64Dayhoff,
+            HashFunctions::Murmur64Hp => Murmur64Hp,
+        }
+    }
+}
+
+impl From<crate::encodings::HashFunctions> for HashFunctions {
+    fn from(v: crate::encodings::HashFunctions) -> HashFunctions {
+        use crate::encodings::HashFunctions::{
+            Murmur64Dayhoff, Murmur64Dna, Murmur64Hp, Murmur64Protein,
+        };
+        match v {
+            Murmur64Dna => HashFunctions::Murmur64Dna,
+            Murmur64Protein => HashFunctions::Murmur64Protein,
+            Murmur64Dayhoff => HashFunctions::Murmur64Dayhoff,
+            Murmur64Hp => HashFunctions::Murmur64Hp,
+            _ => todo!("Not supported, probably custom"),
+        }
+    }
 }
