@@ -29,7 +29,7 @@ pub unsafe extern "C" fn nodegraph_buffer_free(ptr: *mut u8, insize: usize) {
     if ptr.is_null() {
         return;
     }
-    Vec::from_raw_parts(ptr as *mut u8, insize, insize);
+    Vec::from_raw_parts(ptr, insize, insize);
 }
 
 #[no_mangle]
