@@ -3,41 +3,68 @@
 //! An index organizes signatures to allow for fast similarity search.
 //! Some indices also support containment searches.
 
-pub mod bigsi;
 pub mod linear;
+
+#[cfg(not(target_arch = "wasm32"))]
+#[cfg(feature = "branchwater")]
 pub mod revindex;
-pub mod sbt;
 
 pub mod search;
 
-use std::ops::Deref;
 use std::path::Path;
 
-use once_cell::sync::OnceCell;
+use getset::{CopyGetters, Getters, Setters};
+
 use serde::{Deserialize, Serialize};
 use typed_builder::TypedBuilder;
 
-use crate::errors::ReadDataError;
-use crate::index::sbt::{Node, SBT};
+use crate::encodings::Idx;
 use crate::index::search::{search_minhashes, search_minhashes_containment};
 use crate::prelude::*;
-use crate::signature::SigsTrait;
-use crate::sketch::nodegraph::Nodegraph;
-use crate::sketch::Sketch;
-use crate::storage::{InnerStorage, Storage};
-use crate::Error;
+use crate::Result;
 
-pub type MHBT = SBT<Node<Nodegraph>, Signature>;
+#[derive(TypedBuilder, CopyGetters, Getters, Setters, Serialize, Deserialize, Debug, PartialEq)]
+pub struct GatherResult {
+    #[getset(get_copy = "pub")]
+    intersect_bp: usize,
 
-/* FIXME: bring back after MQF works on macOS and Windows
-use cfg_if::cfg_if;
-cfg_if! {
-    if #[cfg(not(target_arch = "wasm32"))] {
-      use mqf::MQF;
-      pub type MHMT = SBT<Node<MQF>, Signature>;
+    #[getset(get_copy = "pub")]
+    f_orig_query: f64,
+
+    #[getset(get_copy = "pub")]
+    f_match: f64,
+
+    f_unique_to_query: f64,
+    f_unique_weighted: f64,
+    average_abund: usize,
+    median_abund: usize,
+    std_abund: usize,
+
+    #[getset(get = "pub")]
+    filename: String,
+
+    #[getset(get = "pub")]
+    name: String,
+
+    #[getset(get = "pub")]
+    md5: String,
+
+    #[serde(skip)]
+    match_: Signature,
+
+    f_match_orig: f64,
+    unique_intersect_bp: usize,
+    gather_result_rank: usize,
+    remaining_bp: usize,
+}
+
+impl GatherResult {
+    pub fn get_match(&self) -> Signature {
+        self.match_.clone()
     }
 }
-*/
+
+type SigCounter = counter::Counter<Idx>;
 
 pub trait Index<'a> {
     type Item: Comparable<Self::Item>;
@@ -48,7 +75,7 @@ pub trait Index<'a> {
         search_fn: F,
         sig: &'a Self::Item,
         threshold: f64,
-    ) -> Result<Vec<&Self::Item>, Error>
+    ) -> Result<Vec<&Self::Item>>
     where
         F: Fn(&dyn Comparable<Self::Item>, &Self::Item, f64) -> bool,
     {
@@ -70,7 +97,7 @@ pub trait Index<'a> {
         sig: &'a Self::Item,
         threshold: f64,
         containment: bool,
-    ) -> Result<Vec<&Self::Item>, Error> {
+    ) -> Result<Vec<&Self::Item>> {
         if containment {
             self.find(search_minhashes_containment, sig, threshold)
         } else {
@@ -78,11 +105,11 @@ pub trait Index<'a> {
         }
     }
 
-    //fn gather(&self, sig: &Self::Item, threshold: f64) -> Result<Vec<&Self::Item>, Error>;
+    //fn gather(&self, sig: &Self::Item, threshold: f64) -> Result<Vec<&Self::Item>>;
 
-    fn insert(&mut self, node: Self::Item) -> Result<(), Error>;
+    fn insert(&mut self, node: Self::Item) -> Result<()>;
 
-    fn batch_insert(&mut self, nodes: Vec<Self::Item>) -> Result<(), Error> {
+    fn batch_insert(&mut self, nodes: Vec<Self::Item>) -> Result<()> {
         for node in nodes {
             self.insert(node)?;
         }
@@ -90,9 +117,9 @@ pub trait Index<'a> {
         Ok(())
     }
 
-    fn save<P: AsRef<Path>>(&self, path: P) -> Result<(), Error>;
+    fn save<P: AsRef<Path>>(&self, path: P) -> Result<()>;
 
-    fn load<P: AsRef<Path>>(path: P) -> Result<(), Error>;
+    fn load<P: AsRef<Path>>(path: P) -> Result<()>;
 
     fn signatures(&self) -> Vec<Self::Item>;
 
@@ -121,214 +148,5 @@ where
 
     fn containment(&self, other: &L) -> f64 {
         (*self).containment(other)
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct DatasetInfo {
-    pub filename: String,
-    pub name: String,
-    pub metadata: String,
-}
-
-#[derive(TypedBuilder, Default, Clone)]
-pub struct SigStore<T> {
-    #[builder(setter(into))]
-    filename: String,
-
-    #[builder(setter(into))]
-    name: String,
-
-    #[builder(setter(into))]
-    metadata: String,
-
-    storage: Option<InnerStorage>,
-
-    #[builder(setter(into), default)]
-    data: OnceCell<T>,
-}
-
-impl<T> SigStore<T> {
-    pub fn name(&self) -> String {
-        self.name.clone()
-    }
-}
-
-impl<T> std::fmt::Debug for SigStore<T> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "SigStore [filename: {}, name: {}, metadata: {}]",
-            self.filename, self.name, self.metadata
-        )
-    }
-}
-
-impl ReadData<Signature> for SigStore<Signature> {
-    fn data(&self) -> Result<&Signature, Error> {
-        if let Some(sig) = self.data.get() {
-            Ok(sig)
-        } else if let Some(storage) = &self.storage {
-            let sig = self.data.get_or_init(|| {
-                let raw = storage.load(&self.filename).unwrap();
-                let sigs: Result<Vec<Signature>, _> = serde_json::from_reader(&mut &raw[..]);
-                if let Ok(sigs) = sigs {
-                    // TODO: select the right sig?
-                    sigs[0].to_owned()
-                } else {
-                    let sig: Signature = serde_json::from_reader(&mut &raw[..]).unwrap();
-                    sig
-                }
-            });
-
-            Ok(sig)
-        } else {
-            Err(ReadDataError::LoadError.into())
-        }
-    }
-}
-
-impl SigStore<Signature> {
-    pub fn count_common(&self, other: &SigStore<Signature>) -> u64 {
-        let ng: &Signature = self.data().unwrap();
-        let ong: &Signature = other.data().unwrap();
-
-        // TODO: select the right signatures...
-        // TODO: better matching here, what if it is not a mh?
-        if let Sketch::MinHash(mh) = &ng.signatures[0] {
-            if let Sketch::MinHash(omh) = &ong.signatures[0] {
-                return mh.count_common(omh, false).unwrap();
-            }
-        }
-        unimplemented!();
-    }
-
-    pub fn mins(&self) -> Vec<u64> {
-        let ng: &Signature = self.data().unwrap();
-
-        // TODO: select the right signatures...
-        // TODO: better matching here, what if it is not a mh?
-        if let Sketch::MinHash(mh) = &ng.signatures[0] {
-            mh.mins()
-        } else {
-            unimplemented!()
-        }
-    }
-}
-
-impl From<SigStore<Signature>> for Signature {
-    fn from(other: SigStore<Signature>) -> Signature {
-        other.data.get().unwrap().to_owned()
-    }
-}
-
-impl Deref for SigStore<Signature> {
-    type Target = Signature;
-
-    fn deref(&self) -> &Signature {
-        self.data.get().unwrap()
-    }
-}
-
-impl From<Signature> for SigStore<Signature> {
-    fn from(other: Signature) -> SigStore<Signature> {
-        let name = other.name();
-        let filename = other.filename();
-
-        SigStore::builder()
-            .name(name)
-            .filename(filename)
-            .data(other)
-            .metadata("")
-            .storage(None)
-            .build()
-    }
-}
-
-impl Comparable<SigStore<Signature>> for SigStore<Signature> {
-    fn similarity(&self, other: &SigStore<Signature>) -> f64 {
-        let ng: &Signature = self.data().unwrap();
-        let ong: &Signature = other.data().unwrap();
-
-        // TODO: select the right signatures...
-        // TODO: better matching here, what if it is not a mh?
-        if let Sketch::MinHash(mh) = &ng.signatures[0] {
-            if let Sketch::MinHash(omh) = &ong.signatures[0] {
-                return mh.similarity(omh, true, false).unwrap();
-            }
-        }
-
-        /* FIXME: bring back after boomphf changes
-        if let Sketch::UKHS(mh) = &ng.signatures[0] {
-            if let Sketch::UKHS(omh) = &ong.signatures[0] {
-                return 1. - mh.distance(&omh);
-            }
-        }
-        */
-
-        unimplemented!()
-    }
-
-    fn containment(&self, other: &SigStore<Signature>) -> f64 {
-        let ng: &Signature = self.data().unwrap();
-        let ong: &Signature = other.data().unwrap();
-
-        // TODO: select the right signatures...
-        // TODO: better matching here, what if it is not a mh?
-        if let Sketch::MinHash(mh) = &ng.signatures[0] {
-            if let Sketch::MinHash(omh) = &ong.signatures[0] {
-                let common = mh.count_common(omh, false).unwrap();
-                let size = mh.size();
-                return common as f64 / size as f64;
-            }
-        }
-        unimplemented!()
-    }
-}
-
-impl Comparable<Signature> for Signature {
-    fn similarity(&self, other: &Signature) -> f64 {
-        // TODO: select the right signatures...
-        // TODO: better matching here, what if it is not a mh?
-        if let Sketch::MinHash(mh) = &self.signatures[0] {
-            if let Sketch::MinHash(omh) = &other.signatures[0] {
-                return mh.similarity(omh, true, false).unwrap();
-            }
-        }
-
-        /* FIXME: bring back after boomphf changes
-        if let Sketch::UKHS(mh) = &self.signatures[0] {
-            if let Sketch::UKHS(omh) = &other.signatures[0] {
-                return 1. - mh.distance(&omh);
-            }
-        }
-        */
-
-        unimplemented!()
-    }
-
-    fn containment(&self, other: &Signature) -> f64 {
-        // TODO: select the right signatures...
-        // TODO: better matching here, what if it is not a mh?
-        if let Sketch::MinHash(mh) = &self.signatures[0] {
-            if let Sketch::MinHash(omh) = &other.signatures[0] {
-                let common = mh.count_common(omh, false).unwrap();
-                let size = mh.size();
-                return common as f64 / size as f64;
-            }
-        }
-        unimplemented!()
-    }
-}
-
-impl<L> From<DatasetInfo> for SigStore<L> {
-    fn from(other: DatasetInfo) -> SigStore<L> {
-        SigStore {
-            filename: other.filename,
-            name: other.name,
-            metadata: other.metadata,
-            storage: None,
-            data: OnceCell::new(),
-        }
     }
 }
