@@ -2,6 +2,8 @@
 //!
 //! A signature is a collection of sketches for a genomic dataset.
 
+use core::iter::FusedIterator;
+
 use std::fs::File;
 use std::io;
 use std::iter::Iterator;
@@ -16,10 +18,13 @@ use typed_builder::TypedBuilder;
 
 use crate::encodings::{aa_to_dayhoff, aa_to_hp, revcomp, to_aa, HashFunctions, VALID};
 use crate::prelude::*;
+use crate::selection::{Select, Selection};
 use crate::sketch::Sketch;
 use crate::Error;
 use crate::HashIntoType;
 
+// TODO: this is the behavior expected from Sketch, but that name is already
+// used. Sketchable?
 pub trait SigsTrait {
     fn size(&self) -> usize;
     fn to_vec(&self) -> Vec<u64>;
@@ -366,11 +371,11 @@ impl Iterator for SeqToHashes {
                     Some(Ok(hash))
                 } else {
                     if !self.prot_configured {
-                        self.aa_seq = match self.hash_function {
-                            HashFunctions::murmur64_dayhoff => {
+                        self.aa_seq = match &self.hash_function {
+                            HashFunctions::Murmur64Dayhoff => {
                                 self.sequence.iter().cloned().map(aa_to_dayhoff).collect()
                             }
-                            HashFunctions::murmur64_hp => {
+                            HashFunctions::Murmur64Hp => {
                                 self.sequence.iter().cloned().map(aa_to_hp).collect()
                             }
                             invalid => {
@@ -395,6 +400,10 @@ impl Iterator for SeqToHashes {
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, TypedBuilder)]
+#[cfg_attr(
+    feature = "rkyv",
+    derive(rkyv::Serialize, rkyv::Deserialize, rkyv::Archive)
+)]
 pub struct Signature {
     #[serde(default = "default_class")]
     #[builder(default = default_class())]
@@ -575,9 +584,9 @@ impl Signature {
                                 }
                             };
 
-                            match moltype {
+                            match &moltype {
                                 Some(x) => {
-                                    if mh.hash_function() == x {
+                                    if mh.hash_function() == *x {
                                         return true;
                                     }
                                 }
@@ -591,9 +600,9 @@ impl Signature {
                                 }
                             };
 
-                            match moltype {
+                            match &moltype {
                                 Some(x) => {
-                                    if mh.hash_function() == x {
+                                    if mh.hash_function() == *x {
                                         return true;
                                     }
                                 }
@@ -654,6 +663,92 @@ impl Signature {
 
         Ok(())
     }
+
+    pub fn iter_mut(&mut self) -> IterMut<'_> {
+        let length = self.signatures.len();
+        IterMut {
+            iter: self.signatures.iter_mut(),
+            length,
+        }
+    }
+
+    pub fn iter(&self) -> Iter<'_> {
+        let length = self.signatures.len();
+        Iter {
+            iter: self.signatures.iter(),
+            length,
+        }
+    }
+}
+
+pub struct IterMut<'a> {
+    iter: std::slice::IterMut<'a, Sketch>,
+    length: usize,
+}
+
+impl<'a> IntoIterator for &'a mut Signature {
+    type Item = &'a mut Sketch;
+    type IntoIter = IterMut<'a>;
+
+    fn into_iter(self) -> IterMut<'a> {
+        self.iter_mut()
+    }
+}
+
+impl<'a> Iterator for IterMut<'a> {
+    type Item = &'a mut Sketch;
+
+    fn next(&mut self) -> Option<&'a mut Sketch> {
+        if self.length == 0 {
+            None
+        } else {
+            self.length -= 1;
+            self.iter.next()
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.length, Some(self.length))
+    }
+}
+
+pub struct Iter<'a> {
+    iter: std::slice::Iter<'a, Sketch>,
+    length: usize,
+}
+
+impl<'a> Iterator for Iter<'a> {
+    type Item = &'a Sketch;
+
+    fn next(&mut self) -> Option<&'a Sketch> {
+        if self.length == 0 {
+            None
+        } else {
+            self.length -= 1;
+            self.iter.next()
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.length, Some(self.length))
+    }
+}
+
+impl FusedIterator for Iter<'_> {}
+
+impl ExactSizeIterator for Iter<'_> {
+    fn len(&self) -> usize {
+        self.length
+    }
+}
+
+impl Clone for Iter<'_> {
+    fn clone(&self) -> Self {
+        Iter {
+            iter: self.iter.clone(),
+            length: self.length,
+        }
+    }
 }
 
 impl ToWriter for Signature {
@@ -663,6 +758,36 @@ impl ToWriter for Signature {
     {
         serde_json::to_writer(writer, &vec![&self])?;
         Ok(())
+    }
+}
+
+impl Select for Signature {
+    fn select(mut self, selection: &Selection) -> Result<Self, Error> {
+        self.signatures.retain(|s| {
+            let mut valid = true;
+            valid = if let Some(ksize) = selection.ksize() {
+                let k = s.ksize() as u32;
+                k == ksize || k == ksize * 3
+            } else {
+                valid
+            };
+            // TODO: execute downsample if needed
+
+            /*
+            valid = if let Some(abund) = selection.abund() {
+                valid && *s.with_abundance() == abund
+            } else {
+                valid
+            };
+            valid = if let Some(moltype) = selection.moltype() {
+                valid && s.moltype() == moltype
+            } else {
+                valid
+            };
+            */
+            valid
+        });
+        Ok(self)
     }
 }
 
