@@ -1,2 +1,2 @@
-from .__main__ import main
+from .__main__ import *         # bring all functions into top-level
 from . import grep
