@@ -45,6 +45,15 @@ def test_run_sourmash_sig_cmd():
     assert status != 0                    # no args provided, ok ;)
 
 
+def test_run_cat_via_parse_args():
+    # run a command ('sourmash.sig.cat') with args constructed via parse_args
+    import sourmash.sig, sourmash.cli
+    sig47 = utils.get_test_data('47.fa.sig')
+
+    args = sourmash.cli.parse_args(['sig', 'cat', sig47])
+    sourmash.sig.cat(args)
+
+
 def test_sig_merge_1_use_full_signature_in_cmd(runtmp):
     c = runtmp
 
@@ -1456,6 +1465,42 @@ def test_sig_split_3_multisig(c):
         assert os.path.exists(c.output(filename))
 
 
+def test_sig_split_3_multisig_sig_gz(runtmp):
+    # split 47 and 47+63-multisig.sig with a .sig.gz extension
+    c = runtmp
+
+    sig47 = utils.get_test_data('47.fa.sig')
+    multisig = utils.get_test_data('47+63-multisig.sig')
+    c.run_sourmash('sig', 'split', sig47, multisig, '-E', '.sig.gz')
+
+    outlist = ['57e2b22f.k=31.scaled=1000.DNA.dup=0.none.sig.gz',
+               'bde81a41.k=31.scaled=1000.DNA.dup=0.none.sig.gz',
+               'f033bbd8.k=31.scaled=1000.DNA.dup=0.none.sig.gz',
+               '87a9aec4.k=31.scaled=1000.DNA.dup=0.none.sig.gz',
+               '837bf2a7.k=31.scaled=1000.DNA.dup=0.none.sig.gz',
+               '485c3377.k=31.scaled=1000.DNA.dup=0.none.sig.gz']
+    for filename in outlist:
+        assert os.path.exists(c.output(filename))
+
+
+def test_sig_split_3_multisig_zip(runtmp):
+    # split 47 and 47+63-multisig.sig with a .zip extension
+    c = runtmp
+
+    sig47 = utils.get_test_data('47.fa.sig')
+    multisig = utils.get_test_data('47+63-multisig.sig')
+    c.run_sourmash('sig', 'split', sig47, multisig, '-E', '.zip')
+
+    outlist = ['57e2b22f.k=31.scaled=1000.DNA.dup=0.none.zip',
+               'bde81a41.k=31.scaled=1000.DNA.dup=0.none.zip',
+               'f033bbd8.k=31.scaled=1000.DNA.dup=0.none.zip',
+               '87a9aec4.k=31.scaled=1000.DNA.dup=0.none.zip',
+               '837bf2a7.k=31.scaled=1000.DNA.dup=0.none.zip',
+               '485c3377.k=31.scaled=1000.DNA.dup=0.none.zip']
+    for filename in outlist:
+        assert os.path.exists(c.output(filename))
+
+
 @utils.in_tempdir
 def test_sig_split_4_sbt_prot(c):
     # split sbt
@@ -2644,7 +2689,7 @@ def test_sig_extract_12_picklist_bad_pickstyle(runtmp):
 
     err = runtmp.last_result.err
     print(err)
-    assert "invalid picklist 'pickstyle' argument, 'XXX': must be 'include' or 'exclude'" in err
+    assert "invalid picklist 'pickstyle' argument 4: 'XXX' must be 'include' or 'exclude'" in err
 
 
 def test_sig_extract_12_picklist_bad_colname(runtmp):
@@ -2717,6 +2762,31 @@ def test_sig_extract_11_pattern_exclude(runtmp):
     names = [ ss.name for ss in idx.signatures() ]
     for n in names:
         assert 'shewanella' not in n.lower(), n
+
+
+def test_sig_extract_identical_md5s(runtmp):
+    # test that we properly handle different signatures with identical md5s
+    sig47 = utils.get_test_data('47.fa.sig')
+    ss = load_signatures(sig47)
+    sig = list(ss)[0]
+    new_sig = sig.to_mutable()
+    new_sig.name = 'foo'
+    sig47foo = runtmp.output('foo.sig')
+    # this was only a problem when the signatures are stored in the same file
+    with open(sig47foo, 'wt') as fp:
+        sourmash.save_signatures([new_sig, sig], fp)
+
+    runtmp.run_sourmash('sig', 'extract', '--name', 'foo', sig47foo)
+
+    out = runtmp.last_result.out
+    print(out)
+    ss = load_signatures(out)
+    ss = list(ss)
+    assert len(ss) == 1
+    ss = ss[0]
+    assert 'Shewanella' not in ss.name
+    assert 'foo' in ss.name
+    assert ss.md5sum() == '09a08691ce52952152f0e866a59f6261'
 
 
 def test_sig_flatten_1(runtmp):
@@ -3924,7 +3994,8 @@ def test_sig_kmers_1_dna(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 1
     assert len(records[0].sequence) == 1000, len(records[0].sequence)
 
@@ -3961,8 +4032,9 @@ def test_sig_kmers_1_dna_more_in_query(runtmp):
     # make a new sequence for query, with more k-mers
     query_seqfile = runtmp.output('query.fa')
     with open(query_seqfile, 'wt') as fp:
-        for record in screed.open(seqfile):
-            fp.write(f">{record.name}\n{record.sequence}AGTTACGATC\n")
+        with screed.open(seqfile) as screed_iter:
+            for record in screed_iter:
+                fp.write(f">{record.name}\n{record.sequence}AGTTACGATC\n")
 
     runtmp.sourmash('sig', 'kmers', '--sig', 'short.fa.sig',
                     '--seq', query_seqfile)
@@ -4070,7 +4142,8 @@ def test_sig_kmers_1_dna_lowscaled(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 1
     assert len(records[0].sequence) == 1000, len(records[0].sequence)
 
@@ -4119,7 +4192,8 @@ def test_sig_kmers_1_dna_num(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 1
     assert len(records[0].sequence) == 1000, len(records[0].sequence)
 
@@ -4168,7 +4242,8 @@ def test_sig_kmers_1_dna_translate_protein(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 1
     assert len(records[0].sequence) == 1000, len(records[0].sequence)
 
@@ -4217,7 +4292,8 @@ def test_sig_kmers_1_dna_translate_dayhoff(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 1
     assert len(records[0].sequence) == 1000, len(records[0].sequence)
 
@@ -4266,7 +4342,8 @@ def test_sig_kmers_1_dna_translate_hp(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 1
     assert len(records[0].sequence) == 1000, len(records[0].sequence)
 
@@ -4315,7 +4392,8 @@ def test_sig_kmers_2_protein(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 2
     assert len(records[0].sequence) == 820, len(records[0].sequence)
     assert len(records[1].sequence) == 310, len(records[1].sequence)
@@ -4365,7 +4443,8 @@ def test_sig_kmers_2_dayhoff(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 2
     assert len(records[0].sequence) == 820, len(records[0].sequence)
     assert len(records[1].sequence) == 310, len(records[1].sequence)
@@ -4415,7 +4494,8 @@ def test_sig_kmers_2_hp(runtmp):
 
     # check FASTA output
     assert os.path.exists(runtmp.output('matched.fa'))
-    records = list(screed.open(runtmp.output('matched.fa')))
+    with screed.open(runtmp.output('matched.fa')) as f:
+        records = list(f)
     assert len(records) == 2
     assert len(records[0].sequence) == 820, len(records[0].sequence)
     assert len(records[1].sequence) == 310, len(records[1].sequence)
